move packages to notify
